Move Status into separate unit
